Consistent page titles
As a rule, you want to use the same page title in the `<title>` and the `<h1>` element. The only difference is that in the `<title>` you also want to include the site name so that Screen readers have context. As with most things related to screen readers, you want to put the page title in front of the name of the site, so that the more important information is read before the generic information. We are also sending page titles to Adobe Analytics as well, so I've used the same as what the `<title>` is to send to AA.

To take the "About these labs" page as an example:

- `<title>`: `About these labs — Service Canada Labs`
- Header meta (“dcterms.title”) for Adobe Analytics: `About these labs — Service Canada Labs`
- `<h1>`: `About these labs`

Since we want the "h1" content to match the "title", I've also removed a bunch of our specific "scLabsAboutTitle" keys because we don't use them anymore.