Replace HTMList in projects pages with imported component
While I was working on the Digital Centre page, I created an HTMList element to make creating lists more efficient, and I left that code in pages/projects/digital-center.js file until now. It was a useful bit of code, so @krischarbonneau picked it up for his file, creating a duplicated component in pages/projects/life-journeys.js. It was on my list to clean this up, and this PR does that.